moved color functions to new logging.go file and updated log formats
