+++ commands/init.go
@@ -1,101 +1,55 @@
 package commands
 
-// Init & Logging sit here
+// Init sits here
 
-import (
-	log "github.com/Sirupsen/logrus"
-	"github.com/spf13/cobra"
-	prefixed "github.com/x-cray/logrus-prefixed-formatter"
-)
-
-// Used for mapping log level, may or may not expand in the future..
-var level = struct {
-	debug string
-	warn  string
-	err   string
-	info  string
-}{"debug", "warn", "error", "info"}
-
-// handleError - handleError logs the err and exits the app if err not nil
-func handleError(e error) {
-	if e != nil {
-		log.WithFields(log.Fields{
-			"request": job.request,
-		}).Errorln(e.Error()) // logrus error calls os.exit after logging
-	}
-}
-
-// Log - Handles all logging accross app
-func Log(msg, lvl string) {
-
-	// TODO: Get this to only check once.
-	if job.debug {
-		log.SetLevel(log.DebugLevel)
-	} else {
-		log.SetLevel(log.InfoLevel)
-	}
-
-	l := log.WithFields(log.Fields{
-		"request": job.request,
-	})
-
-	switch lvl {
-	case "debug":
-		l.Debugln(msg)
-	case "warn":
-		l.Warnln(msg)
-	case "error":
-		l.Errorln(msg)
-	default:
-		l.Infoln(msg)
-	}
-}
+import "github.com/spf13/cobra"
 
 func init() {
 
 	// Define Deploy Flags
-	deployCmd.Flags().StringArrayVarP(&job.tplSources, "template", "t", []string{}, "path to template file(s) Or stack::url")
-	deployCmd.Flags().BoolVarP(&job.rollback, "rollback", "R", false, "Set Stack to rollback on deployment failures")
-	deployCmd.Flags().BoolVarP(&job.all, "all", "A", false, "deploy all stacks with defined Sources in config")
+	deployCmd.Flags().StringArrayVarP(&run.tplSources, "template", "t", []string{}, "path to template file(s) Or stack::url")
+	deployCmd.Flags().BoolVarP(&run.rollback, "rollback", "R", false, "Set Stack to rollback on deployment failures")
+	deployCmd.Flags().BoolVarP(&run.all, "all", "A", false, "deploy all stacks with defined Sources in config")
 
 	// Define Terminate Flags
-	terminateCmd.Flags().BoolVarP(&job.all, "all", "A", false, "terminate all stacks")
+	terminateCmd.Flags().BoolVarP(&run.all, "all", "A", false, "terminate all stacks")
 
 	// Define Output Flags
-	outputsCmd.Flags().StringVarP(&job.profile, "profile", "p", "default", "configured aws profile")
+	outputsCmd.Flags().StringVarP(&run.profile, "profile", "p", "default", "configured aws profile")
 
 	// Define Exports Flags
 	exportsCmd.Flags().StringVarP(&region, "region", "r", "eu-west-1", "AWS Region")
 
 	// Define Root Flags
-	RootCmd.Flags().BoolVarP(&job.version, "version", "", false, "print current/running version")
-	RootCmd.PersistentFlags().StringVarP(&job.profile, "profile", "p", "default", "configured aws profile")
-	RootCmd.PersistentFlags().BoolVarP(&job.debug, "debug", "", false, "Run in debug mode...")
+	RootCmd.Flags().BoolVarP(&run.version, "version", "", false, "print current/running version")
+	RootCmd.PersistentFlags().BoolVarP(&run.colors, "no-colors", "", false, "disable colors in outputs")
+	RootCmd.PersistentFlags().StringVarP(&run.profile, "profile", "p", "default", "configured aws profile")
+	RootCmd.PersistentFlags().BoolVarP(&run.debug, "debug", "", false, "Run in debug mode...")
 
 	// Define Invoke Flags
 	invokeCmd.Flags().StringVarP(&region, "region", "r", "eu-west-1", "AWS Region")
-	invokeCmd.Flags().StringVarP(&job.funcEvent, "event", "e", "", "JSON Event data for AWS Lambda invoke")
+	invokeCmd.Flags().StringVarP(&run.funcEvent, "event", "e", "", "JSON Event data for AWS Lambda invoke")
 
 	// Define Changes Command
 	changeCmd.AddCommand(create, rm, list, execute, desc)
 
 	// Add Config --config common flag
 	for _, cmd := range []interface{}{checkCmd, updateCmd, outputsCmd, statusCmd, terminateCmd, generateCmd, deployCmd, policyCmd} {
-		cmd.(*cobra.Command).Flags().StringVarP(&job.cfgSource, "config", "c", defaultConfig(), "path to config file")
+		cmd.(*cobra.Command).Flags().StringVarP(&run.cfgSource, "config", "c", defaultConfig(), "path to config file")
 	}
 
 	// Add Template --template common flag
 	for _, cmd := range []interface{}{generateCmd, updateCmd, checkCmd} {
-		cmd.(*cobra.Command).Flags().StringVarP(&job.tplSource, "template", "t", "", "path to template source Or stack::source")
+		cmd.(*cobra.Command).Flags().StringVarP(&run.tplSource, "template", "t", "", "path to template source Or stack::source")
 	}
 
 	for _, cmd := range []interface{}{create, list, rm, execute, desc} {
-		cmd.(*cobra.Command).Flags().StringVarP(&job.cfgSource, "config", "c", defaultConfig(), "path to config file [Required]")
-		cmd.(*cobra.Command).Flags().StringVarP(&job.stackName, "stack", "s", "", "Qaz local project Stack Name [Required]")
+		cmd.(*cobra.Command).Flags().StringVarP(&run.cfgSource, "config", "c", defaultConfig(), "path to config file [Required]")
+		cmd.(*cobra.Command).Flags().StringVarP(&run.stackName, "stack", "s", "", "Qaz local project Stack Name [Required]")
 	}
 
-	create.Flags().StringVarP(&job.tplSource, "template", "t", "", "path to template file Or stack::url")
-	changeCmd.Flags().StringVarP(&job.cfgSource, "config", "c", defaultConfig(), "path to config file")
+	create.Flags().StringVarP(&run.tplSource, "template", "t", "", "path to template file Or stack::url")
+	changeCmd.Flags().StringVarP(&run.cfgSource, "config", "c", defaultConfig(), "path to config file")
 
 	RootCmd.AddCommand(
 		generateCmd, deployCmd, terminateCmd,
@@ -104,6 +58,4 @@ func init() {
 		invokeCmd, changeCmd, policyCmd,
 	)
 
-	// Setup logging
-	log.SetFormatter(new(prefixed.TextFormatter))
 }
